backend/pyproject.toml
[project]
name = "thela-backend"
version = "0.1.0"
description = "Thela FastAPI backend replacing Supabase with Google OAuth and S3 support"
requires-python = ">=3.11"
authors = [{name = "Thela", email = "dev@thela.app"}]
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.111,<1.0",
    "uvicorn[standard]>=0.30,<0.31",
    "sqlalchemy>=2.0,<3.0",
    "asyncpg>=0.29,<0.30",
    "alembic>=1.13,<1.14",
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "google-auth>=2.30,<3.0",
    "pyjwt>=2.8,<3.0",
    "python-jose[cryptography]>=3.3,<4.0",
    "passlib[bcrypt]>=1.7,<2.0",
    "python-multipart>=0.0.9,<0.1",
    "boto3>=1.34,<2.0",
    "slowapi>=0.1.9,<0.2",
    "meilisearch>=0.31,<1.0",
    "httpx>=0.27,<0.28"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2,<9.0",
    "pytest-asyncio>=0.23,<0.24",
    "httpx>=0.27,<0.28",
    "ruff>=0.5,<0.6"
]

[project.urls]
Homepage = "https://thela.app"

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
